Create Duplicate Information from One Sheet onto Another with Condition in Column
Very simply will check the main sheet's A:A for the string "LEFT", and if true, will copy columns A:I over. To snapshot this information in the current version of Excel Online, either use Power Automate, or manually copy A:I, and paste the values as text (ctrl+shift+v) over the top of it.