Create Fuzzy Extract Dates from PDF File Names From Reference
Full Power Query code for adding a helper column that attempts to extract a date from a file's name. If this is not possible it will use date the file was last modified using the query 'FilesMeta'.

This is cleanest implemented manually step-by-step in Power Query: After 'Source' (which should be a reference link to your dataset query), the steps 'AddedFromName', 'MergedMeta', 'ExpandedMeta', and 'AddedExportDate' should be made and each line inserted in the Advanced Editor.