Create Find Duplicate Information from Contact List with Multiple Contacts per CustomerID
This is a basic outline of what queries can be used to identify information that exists on multiple rows over any number of identifiable data.

In this example, it groups and filters out a customer list based on whether the correspondence method was set to "Post". It then checks if all contacts for that customer are set to "Post" and only shows those.

This is useful if this data would be an error if "Post" is a correspondence method for secondary contacts but is applied to all contacts. Or for matching against a contact list so that all correspondence can go to the CustomerID instead of line-by-line checking contacts for that CustomerID if the messaging method allows.