Create Neatly printing text using Layers of Filters
This will print whatever is in Sheet2's column 'C' if:
- Column A in Sheet1 matches column A in Sheet2
and
- Column B's contents in Sheet2 match column B in Sheet1

This is particularly useful if you have information in 'C' but need to make sure it fits exactly with another element, so instead of, say, printing it to a helper column then using the helper column to filter, you can just embed the check into 'ISNUMBER' as if it doesn't match it'll be 0 and will error out at the TEXTJOIN level. Avoiding accidentally printing information that could potentially be harmful if exported into another process.